Add files via upload
Transcribed from Scratch

I couldn’t get Audiveris to work, so I ended up doing everything manually in MuseScore Studio. It was definitely a challenge, but once I found the right template and used an existing guitar sheet as a reference, it started to come together. The formatting tools in MuseScore were pretty frustrating at first especially since I’m not very experienced with music software but after spending some time with it and getting used to the interface, it eventually felt as simple as working in Google Docs.